backend/pyproject.toml
[project]
name = "gigame-backend"
version = "0.1.0"
description = "GigaMe — GigaChat-powered document processing platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "gigachat>=0.1.36",
    "python-multipart>=0.0.9",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "sse-starlette>=2.0.0",
    "aiofiles>=24.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
